refactor: extract astra-turn-core with 16 pure turn modules
Move 16 zero-coupling turn modules (~2,400 LOC, 130 tests) from
runtime into the new astra-turn-core crate:

Types & contracts:
- confidence_contract: tier classification and trend tracking
- agentic_verdict_audit: verdict audit event struct
- chat_turn_api_error: HTTP error user messages
- chat_turn_explain_wire: explain UI mode/flags
- followup_suggestion: next-step suggestion engine

Session state:
- state: session entry normalization
- execution_state: execution state normalization
- cache: generic session cache (LRU + TTL)
- snapshot: session history snapshots

Helpers:
- unconsumed: tool message consumption tracking
- tool_call_shape: tool call name/args parsing
- firewall: verification plan builder
- observer: observer message builder
- explain: explain event builder
- task: task classifier
- edge_executor_id: random executor ID generator

Runtime keeps thin re-export layers for full API compatibility.
