+++ rust/crates/astra-turn-core/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "astra-turn-core"
+version.workspace = true
+edition.workspace = true
+
+[dependencies]
+chrono = { workspace = true }
+regex = { workspace = true }
+serde = { workspace = true, features = ["derive"] }
+serde_json = { workspace = true }
+uuid = { workspace = true, features = ["v4"] }
