+++ rust/crates/runtime/Cargo.toml
@@ -36,6 +36,7 @@ astra-pipeline.workspace = true
 astra-sandbox.workspace = true
 astra-skills.workspace = true
 astra-server-types.workspace = true
+astra-turn-core.workspace = true
 astra-turn-types.workspace = true
 astra-services.workspace = true
 astra-tools.workspace = true
